+++ CHANGELOG.md
@@ -10,6 +10,8 @@ KleidiAI follows the [Semantic Versioning](https://semver.org/) specification fo
 
 ## Upcoming Release
 
+- New SME2 QSI8D32P x QSI4C32P dot matrix-vector multiplication micro-kernel with optional 16-entry LUT decoding for blockwise 4-bit formats.
+
 ## v1.30.0
 
 - New SME2 micro-kernels